Fixed open class time in female half-marathon
It was pointed out by Tom Fryers that the pace (seconds/km) for the female half-marathon was less than the pace for the female 20 km even though the half-marathon is slightly longer than 20 km. This has been corrected in this upload by changing the open class time for the female half-marathon in these three files.